ambient: use "CNIMode" (now renamed) for inpod rules (#51777)
"CNIMode" was a mode added to control how IPTables rules are executed
when running like CNI. In particular, that means we are in the host
filesystem, but the pod network namespace. In this case, we would be
fetching the node level lock which doesn't make sense. There is also
some crazy NSS interactions on some platforms; notably, in GKE, this can
lead to ~1min pod startup times.

Another issue with the locking is a bug in some versions of iptables,
including the one in 1.22. In these versions, `--wait` only acts on 1s
intervals. So if I start many pods at the same time, they will take
`NPods seconds` to start. This is observed in real world clusters almost
100% consistently when many cronjobs trigger pods at the same time. By
avoiding the lock, we sidestep this problem.

Because ambient runs in two modes -- in pod, and host -- "CNI Mode" is
kind of misleading. In this PR, I have renamed it to
HostFilesystemPodNetwork to be a bit more clear. Additionally, to handle
these two modes, the ambient server now has two iptables configs: one
for usage in host, and one for usage in pods.

Given the impact on 1.22 with locking, I would like to backport this to
1.22 if feasible.