Improve testing of overlapping CIDRs (#52236)
This test is... weird. I don't think its valid.

It is meant to test overlapping CIDRs, but it is actually testing two
services with the same hostname, so one is entirely ignored.

I have updated the test to test real overlapping IPs. In doing this, I
found another bug when SNI+CIDR is set together (very niche), tracking
in https://github.com/istio/istio/issues/52235. I did NOT fix this bug
in this PR, merely test it.

In order to facilitate this, I made changes to our FCM to match Envoy to
chose the closest matches, not only a boolean of exact matching

Related to https://github.com/istio/istio/pull/51776