ambient: make HostNetwork pods work (#52191)
This contains a series of fixes to make pods in the hostNetwork work.

On the surface, the fix is trivial: just allow hostNetwork pods. The
issue arises in the fact this now allows duplication of IPs, since
multiple host network pods on the same node will share the same IP.

We address this in a number of ways.

1. Remove the restriction on overlapping IPs that we currently have. If
   the pod is a host network pod, we allow multiple with the same IP.
  1a. In order to facilitate this, we indicate this in WDS. This allows
Ztunnel to know it should not treat the address as unique. In
particular, this means that it will never do an incorrect lookup when
calling the pod directly. Without this change, we may see `curl
host-network-pod`, ztunnel look's up the IP, and finds
`some-other-host-network-pod`. With this change, we recognize these IPs
are not valid to lookup like this and skip them. See the Ztunnel PR for
more info.
2. Fix our EndpointSlice lookup logic. Today, its based on IP matching.
   This isn't really safe in general, but especially not with our new
world of duplicate IPs. Switch this logic to specifically lookup by the
targetRef, which is more clear, safer, and more efficient.

Some minor changes are commented inline in the code